requirements.txt
streamlit
google_generativeai
python-dotenv